conplete flex-modal
Goodness gracious, that felt like a boss fight. I just stared at
it for a good 30 minutes, doing nothing before I forced myself to
visualize everything as boxes. Got stuck on moving the icon to the left
by itself, but I visualized it as another set of boxes and I got it
going. Phew.